for build.gradle, added implementation
Made a new class, EmailService.java, in order to send email. compatible with current spring boot ver 2.7
** application yaml needs to be edited (which means actual email address and server and port in order to send email)

for PriorifyBackendApplication.java, added @ and corresponding springboot framework import

for ScheduleService.java, I made a method to send email with a example format which later be trimmed.

I tried to sort tasks with days to come, then alphabet order. Meanwhile, I failed to sort with priority-score order. Suggest this method as a prototype later to be added if development schedule and process would allow.